+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added PUSH/PROCESS checkboxes to filter frames by event type.
 
+- Added frame count display showing the number of filtered frames out of the
+  total.
+
 ## [0.0.10] - 2025-10-03
 
 ### Added